requirement.txt
pandas
numpy
requests
beautifulsoup4
lxml
mysql-connector-python
python-dotenv
